+++ src/FluxFlow.Components.Mapping.Composition/MappingCompositionNodeRegistryExtensions.cs
@@ -2,12 +2,35 @@ using FluxFlow.Components.Mapping.Contracts;
 using FluxFlow.Components.Mapping.Nodes;
 using FluxFlow.Components.Mapping.Options;
 using FluxFlow.Composition;
+using FluxFlow.Data;
 using FluxFlow.Mapping;
 
 namespace FluxFlow.Components.Mapping.Composition;
 
 public static class MappingCompositionNodeRegistryExtensions
 {
+    public static CompositionNodeRegistry RegisterMapper(
+        this CompositionNodeRegistry registry,
+        string nodeType = MappingCompositionNodeTypes.Mapper)
+    {
+        ArgumentNullException.ThrowIfNull(registry);
+        ArgumentException.ThrowIfNullOrWhiteSpace(nodeType);
+
+        return registry.Register(
+            nodeType,
+            CreateFlowValueMapperNode,
+            inputs:
+            [
+                CompositionPorts.Metadata<FlowValue>(
+                    MappingCompositionPortNames.Input)
+            ],
+            outputs:
+            [
+                CompositionPorts.Metadata<FlowResult<FlowValue>>(
+                    MappingCompositionPortNames.Output)
+            ]);
+    }
+
     public static CompositionNodeRegistry RegisterMapper<TInput, TOutput>(
         this CompositionNodeRegistry registry,
         string nodeType = MappingCompositionNodeTypes.Mapper)
@@ -30,6 +53,39 @@ public static class MappingCompositionNodeRegistryExtensions
                 CompositionPorts.Metadata<TInput>(
                     MappingCompositionPortNames.Failed)
             ]);
+    }
+
+    private static ValueTask<ComposedNode> CreateFlowValueMapperNode(
+        CompositionNodeFactoryContext context)
+    {
+        var options = context.BindConfiguration<MapperOptions>();
+        var expressionEngine = context.GetRequiredResource<IFlowExpressionEngine>(
+            MappingCompositionResourceNames.Engine);
+        var contextFactory = context.GetResource<IMappingContextFactory>(
+            MappingCompositionResourceNames.ContextFactory);
+        var clock = context.GetResource<TimeProvider>(
+            MappingCompositionResourceNames.Clock);
+        var node = new FlowValueMapperNode(
+            options,
+            expressionEngine,
+            contextFactory,
+            clock);
+
+        return ValueTask.FromResult(ComposedNode.Create(
+            node,
+            inputs:
+            [
+                CompositionPorts.Input<FlowValue>(
+                    MappingCompositionPortNames.Input,
+                    node.Input)
+            ],
+            outputs:
+            [
+                CompositionPorts.Output<FlowResult<FlowValue>>(
+                    MappingCompositionPortNames.Output,
+                    node.Output)
+            ],
+            events: node.Events));
     }
 
     private static ValueTask<ComposedNode> CreateMapperNode<TInput, TOutput>(